Extract playback speed related APIs
